Refactor `GsonJobSpecSerDe` into a reusable `GenericGsonSpecSerDe`.
